Common refactoring (#17)

+++ libs/util/include/compression_sorts/suffix_array.hpp
@@ -8,7 +8,7 @@ namespace CompressionSorts {
 
 std::vector<size_t> BuildSuffixArray(const std::string& /*s*/);
 
-std::vector<int> BuildLCPArray(const std::string& /*s*/, const std::vector<size_t>& /*sa*/);
+std::vector<size_t> BuildLCPArray(const std::string& /*s*/, const std::vector<size_t>& /*sa*/);
 
 std::vector<size_t> GetSuffixArrayGreedyOrder(const std::vector<std::string>& /*data*/);
